Revert "refactoring for cleaner organization"
