+++ examples/cvi-transparent-background/.gitignore
@@ -7,11 +7,6 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
-# Environment variables
-.env
-.env.local
-.env.*.local
-
 node_modules
 dist
 dist-ssr
