+++ examples/cvi-transparent-background/src/api/index.ts
@@ -1,4 +1,2 @@
-export { createConversation } from "./createConversation";
-export { endConversation } from "./endConversation";
-export { createPersona } from "./createPersona";
-export { overwriteConversationContext } from "./overwriteContext";
+export * from "./createConversation";
+export * from "./endConversation";
