--- examples/cvi-transparent-background/src/components/Video.tsx
@@ -1,238 +1,0 @@
-import { useRef, useState, useEffect, useMemo } from "react";
-import { DailyVideo } from "@daily-co/daily-react";
-
-/**
- * WebGL Shader Programs for Chroma Key Effect
- */
-const vertexShaderSource = `
-  attribute vec2 a_position;
-  attribute vec2 a_texCoord;
-  varying vec2 v_texCoord;
-  void main() {
-    gl_Position = vec4(a_position, 0, 1);
-    v_texCoord = vec2(a_texCoord.x, 1.0 - a_texCoord.y); // Flip Y coordinate for WebGL
-  }
-`;
-
-const fragmentShaderSource = `
-  precision mediump float;
-  uniform sampler2D u_image;
-  varying vec2 v_texCoord;
-  uniform vec3 u_keyColor;
-  uniform float u_threshold;
-  void main() {
-    vec4 color = texture2D(u_image, v_texCoord);
-    float diff = length(color.rgb - u_keyColor); // Calculate color difference
-    gl_FragColor = diff < u_threshold ? vec4(0.0) : color; // Make matching pixels transparent
-  }
-`;
-
-/**
- * Helper function to create and compile a WebGL shader
- */
-const initShader = (
-  gl: WebGLRenderingContext,
-  type: number,
-  source: string,
-) => {
-  const shader = gl.createShader(type)!;
-  gl.shaderSource(shader, source);
-  gl.compileShader(shader);
-  return shader;
-};
-
-/**
- * Initializes the WebGL context with necessary buffers, attributes and uniforms
- */
-const initWebGL = (gl: WebGLRenderingContext) => {
-  // Create and link shader program
-  const program = gl.createProgram()!;
-  gl.attachShader(
-    program,
-    initShader(gl, gl.VERTEX_SHADER, vertexShaderSource),
-  );
-  gl.attachShader(
-    program,
-    initShader(gl, gl.FRAGMENT_SHADER, fragmentShaderSource),
-  );
-  gl.linkProgram(program);
-  gl.useProgram(program);
-
-  // Set up vertex positions for a full-screen quad
-  const positionBuffer = gl.createBuffer();
-  gl.bindBuffer(gl.ARRAY_BUFFER, positionBuffer);
-  gl.bufferData(
-    gl.ARRAY_BUFFER,
-    new Float32Array([-1, -1, 1, -1, -1, 1, 1, 1]),
-    gl.STATIC_DRAW,
-  );
-
-  // Set up texture coordinates to map video to quad
-  const texCoordBuffer = gl.createBuffer();
-  gl.bindBuffer(gl.ARRAY_BUFFER, texCoordBuffer);
-  gl.bufferData(
-    gl.ARRAY_BUFFER,
-    new Float32Array([0, 0, 1, 0, 0, 1, 1, 1]),
-    gl.STATIC_DRAW,
-  );
-
-  // Connect position and texture coordinate attributes
-  const positionLocation = gl.getAttribLocation(program, "a_position");
-  const texCoordLocation = gl.getAttribLocation(program, "a_texCoord");
-
-  gl.enableVertexAttribArray(positionLocation);
-  gl.bindBuffer(gl.ARRAY_BUFFER, positionBuffer);
-  gl.vertexAttribPointer(positionLocation, 2, gl.FLOAT, false, 0, 0);
-
-  gl.enableVertexAttribArray(texCoordLocation);
-  gl.bindBuffer(gl.ARRAY_BUFFER, texCoordBuffer);
-  gl.vertexAttribPointer(texCoordLocation, 2, gl.FLOAT, false, 0, 0);
-
-  // Create and configure texture for video frames
-  const texture = gl.createTexture();
-  gl.bindTexture(gl.TEXTURE_2D, texture);
-  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
-  gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
-
-  return {
-    program,
-    texture,
-    imageLocation: gl.getUniformLocation(program, "u_image"),
-    keyColorLocation: gl.getUniformLocation(program, "u_keyColor"),
-    thresholdLocation: gl.getUniformLocation(program, "u_threshold"),
-  };
-};
-
-interface VideoProps {
-  id: string;
-}
-
-export const Video: React.FC<VideoProps> = ({ id }) => {
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const videoRef = useRef<HTMLVideoElement>(null);
-  const [isVideoReady, setIsVideoReady] = useState(false);
-  const glRef = useRef<WebGLRenderingContext | null>(null);
-
-  // Initialize WebGL context and resources when canvas is available
-  const webGLContext = useMemo(() => {
-    const canvas = canvasRef.current;
-    if (canvas) {
-      const gl = canvas.getContext("webgl", {
-        premultipliedAlpha: false,
-        alpha: true,
-      });
-      if (gl) {
-        glRef.current = gl;
-        return initWebGL(gl);
-      }
-    }
-    return null;
-  }, [canvasRef.current]);
-
-  // Monitor video element for when it's ready to play
-  useEffect(() => {
-    const video = videoRef.current;
-    if (video) {
-      const checkVideoReady = () => {
-        if (video.readyState >= HTMLMediaElement.HAVE_CURRENT_DATA) {
-          setIsVideoReady(true);
-          video.removeEventListener("canplay", checkVideoReady);
-        }
-      };
-      video.addEventListener("canplay", checkVideoReady);
-      return () => video.removeEventListener("canplay", checkVideoReady);
-    }
-  }, []);
-
-  // Main render loop that applies the chroma key effect
-  useEffect(() => {
-    if (!isVideoReady || !webGLContext) return;
-
-    const video = videoRef.current;
-    const canvas = canvasRef.current;
-    const gl = glRef.current;
-    if (!video || !canvas || !gl) return;
-
-    const {
-      program,
-      texture,
-      imageLocation,
-      keyColorLocation,
-      thresholdLocation,
-    } = webGLContext;
-
-    let animationFrameId: number;
-    let lastFrameTime = 0;
-    const targetFPS = 30;
-    const frameInterval = 1000 / targetFPS;
-
-    const applyChromaKey = (currentTime: number) => {
-      if (currentTime - lastFrameTime < frameInterval) {
-        animationFrameId = requestAnimationFrame(applyChromaKey);
-        return;
-      }
-
-      lastFrameTime = currentTime;
-
-      if (video.readyState === video.HAVE_ENOUGH_DATA) {
-        canvas.width = video.videoWidth;
-        canvas.height = video.videoHeight;
-        gl.viewport(0, 0, canvas.width, canvas.height);
-
-        gl.bindTexture(gl.TEXTURE_2D, texture);
-        gl.texImage2D(
-          gl.TEXTURE_2D,
-          0,
-          gl.RGBA,
-          gl.RGBA,
-          gl.UNSIGNED_BYTE,
-          video,
-        );
-
-        gl.uniform1i(imageLocation, 0);
-        gl.uniform3f(keyColorLocation, 3 / 255, 255 / 255, 156 / 255);
-        gl.uniform1f(thresholdLocation, 0.4);
-
-        gl.drawArrays(gl.TRIANGLE_STRIP, 0, 4);
-      }
-
-      animationFrameId = requestAnimationFrame(applyChromaKey);
-    };
-
-    applyChromaKey(0);
-
-    return () => {
-      cancelAnimationFrame(animationFrameId);
-      if (gl && program && texture) {
-        gl.deleteProgram(program);
-        gl.deleteTexture(texture);
-      }
-    };
-  }, [isVideoReady, webGLContext]);
-
-  return (
-    <div style={{ height: "800px", position: "relative", aspectRatio: "9 / 16" }}>
-      <DailyVideo
-        sessionId={id}
-        type="video"
-        ref={videoRef}
-        style={{ height: "100%", display: "none" }}
-        automirror
-      />
-      <canvas
-        ref={canvasRef}
-        style={{
-          position: "absolute",
-          top: 0,
-          left: 0,
-          right: 0,
-          bottom: 0,
-          width: "100%",
-          height: "100%",
-        }}
-      />
-    </div>
-  );
-};
